license_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2
python-multipart==0.0.6
gunicorn==21.2.0
